# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de API Whatsapp",
        "description": "Guía para el uso de API REST para realizar consultas. Las rutas requieren autenticación previa, obtenga un token desde la ruta /login e ingréselo en Authorize",
        "version": "0.2.0"
    },
    "servers": [
        {
            "url": "se-carga-en-config.js",
            "description": "API WHATSAPP"
        }
    ],
    "components":{
        "securitySchemes":{
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }            
        }  
    },
    "paths": {
...
        "/enviarmensaje": {
            "post": {
                "tags": [
                    "Enviar Mensaje"
                ],
                "security":[
                    {"bearerAuth": []}
                ],
                "summary": "Se envia el body al numero especificado",
                "operationId": "enviarMensaje",
                "requestBody": {
                    "description": "JSON con numero y body del mensaje",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type":"object",
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "5491167070753"
                                    },
                                    "body": {
                                        "type":"string",
                                        "example": "Mensaje a enviar"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Error al enviar",
                        "content": {}
                    },
                    "200": {
                        "description": "OK mensaje enviado",
                        "content": {}
                    }
                }
            }
        },
        "/enviarmensajexid": {
            "post": {
                "tags": [
                    "Enviar Mensaje"
                ],
                "security":[
                    {"bearerAuth": []}
                ],
                "summary": "Se envia el body al jid especificado",
                "operationId": "enviarMensaje",
                "requestBody": {
                    "description": "JSON con numero y body del mensaje",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type":"object",
                                "properties": {
                                    "jid": {
                                        "type": "string",
                                        "example": "5491167070753@c.us"
                                    },
                                    "body": {
                                        "type":"string",
                                        "example": "Mensaje a enviar"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Error al enviar",
                        "content": {}
                    },
                    "200": {
                        "description": "OK mensaje enviado",
                        "content": {}
                    }
                }
            }
        },
...
    }
}